Fixed timekeeping and time calculations
ESPHome does not have good time support.
Timestamps are float, and there is not enough precision for seconds, so text_sensor is used.
But then I need to use text timestamps in the code.
ESPTime does not support subtracting time.
So I had no choice but to use the tm struct and functions to adjust the time, then use ESPTime to render.
I also had issues with the time jumping 1 hour randomly, and more often in high CPU load. Using ESPTime where possible fixed this. It was either a DST issue or a time zone issue + non thread-safe C functions.